refactor(run): extract context helpers to src/lib/run-context.ts (#552) (#560)
Reduces run.ts from 3064 to 2650 lines (-414 lines, ~13.5%) by moving
logically cohesive context-building helpers into a dedicated module:

- gatherSquadContext: squad/agent context assembly for prompt injection
- loadApprovalInstructions: approval/escalation instruction loading
- loadPostExecution: post-execution template loading with placeholder substitution
- AgentFrontmatter interface + parseAgentFrontmatter: YAML frontmatter parser
- extractMcpServersFromDefinition: MCP server discovery from agent definitions

All functions re-exported from run-context.ts with the same signatures.
Removes readdirSync from run.ts imports (no longer needed there).
Build passes, all 1534 tests pass.
